Store phase information for GT field. Phased information is stored in TileDB in the GT field. Examples: VCF: GT 2/3  TileDB: [ 2, 0, 3 ]. The 0 represents unphased allele VCF: GT 2|3  TileDB: [ 2, 1, 3 ] The length descriptor for the GT field in the GenomicsDB vid map must be set and maintained uniformly for a given array - you cannot switch between phased and unphased arbitrarily.
Two length descriptors:
* PP or Phased_Ploidy - store phase information (recommended)
* P - don't store phase information

Effects:
* Ploidy != GT_field.length() - depends on the length descriptor (PP or
P)
* Must deal with inteleaved GT allele and phase information

Notes:
htsjdk deals with phased genotypes in a weird manner. For example, if an
input VCF has the following GT field:
GT: 0/1|2 - triploid with mixed unphased-phased genotypes
With htsjdk VCFCodec, reading and writing out this data produces:
GT: 0|1|2
With htsjdk BCF2Codec, reading and writing out this data produces:
GT: 0/1/2

This occurs with plain htsjdk (PrintVariantsExample), so it's not really
an issue in GenomicsDB.

+++ src/main/cpp/src/genomicsdb/variant_field_handler.cc
@@ -585,7 +585,7 @@ bool VariantFieldHandler<DataType>::concatenate_field(const Variant& variant, co
 
 template<class DataType>
 bool VariantFieldHandler<DataType>::collect_and_extend_fields(const Variant& variant, const VariantQueryConfig& query_config, 
-        unsigned query_idx, const void ** output_ptr, unsigned& num_elements,
+        unsigned query_idx, const void ** output_ptr, uint64_t& num_elements,
         const bool use_missing_values_only_not_vector_end, const bool use_vector_end_only,
         const bool is_GT_field)
 {
